Fixes issues found in Virtual Summit presentation recording
Re-publishing an artifact to zpm-registry would fail because it couldn't be re-published to the local cache. This is valid in some use cases and should be allowed (specifically if changing underlying dependencies such as zpm version impacts the artifact that is produced).

Also fixed an issue installing from local cache with the -verbose flag.