fix: skip tests if not configured
Registry-dependent tests are now skipped if there's no registry.
The integration test that uses simple-app was broken probably by some refactor of %ZHSLIB.AbstractInstaller-related things - added a stub installer class but we'll need to circle back to discussion of having such an interface defined in IPM itself because installer classes are important.