Remove an extra _catalog api call that is made during install against an oras repo
Without the _catalog call, it is already enough to do all necessary filtering for oras package installs. Doing the _catalog call is unnecessary. As _catalog is also not part of oci specs, per https://github.com/opencontainers/distribution-spec/issues/22 , the return behavior is also undefined, which can cause the install commands to fail depending on the specific implementation of the OCI registry, such as in JFrog Artifactory when combined with namespaces.

Signed-off-by: Eric Chen <echen@intersystems.com>